Uploaded "Daily Coding Problem: Problem #1772"
(Traverse the number line from 0 to N by incrementally jumping forward or backward. Report the shortest sequence.)
"Starting from 0 on a number line, you would like to make a series of jumps that lead to the integer N. On the ith jump, you may move exactly i places to the left or right. Find a path with the fewest number of jumps required to get from 0 to N."